- added Orientation enum - small refactoring
git-svn-id: https://jlib-api.svn.sourceforge.net/svnroot/jlib-api/jlib.core/trunk@51 88daaaa0-8d46-0410-88b0-ad3e7f9b0d50

+++ src/org/jlib/core/util/string/StringUtility.java
@@ -114,7 +114,7 @@ public class StringUtility {
                     stringBuffer.append(character);
                 break;
             case FRONTBACK:
-                halfLength = (int) (currentLength + length + 1) / 2;
+                halfLength = (currentLength + length + 1) / 2;
                 for (; currentLength < halfLength; currentLength ++)
                     stringBuffer.append(character);
                 stringBuffer.append(string);
@@ -122,7 +122,7 @@ public class StringUtility {
                     stringBuffer.append(character);
                 break;
             case BACKFRONT:
-                halfLength = (int) (currentLength + length) / 2;
+                halfLength = (currentLength + length) / 2;
                 for (; currentLength < halfLength; currentLength ++)
                     stringBuffer.append(character);
                 stringBuffer.append(string);
